+++ koin-projects/examples/androidx-compose/src/main/java/org/koin/sample/androidx/compose/AppApplication.kt
@@ -3,7 +3,7 @@ package org.koin.sample.androidx.compose
 import android.app.Application
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
-import org.koin.core.context.GlobalContext.startKoin
+import org.koin.core.context.startKoin
 import org.koin.sample.androidx.compose.di.appModule
 
 class AppApplication : Application() {